+++ Expanded Cheat Menu/Code/ReplacedFunctions.lua
@@ -28,26 +28,6 @@ local function SetTrans(dlg)
 	return dlg
 end
 
-do -- do some stuff
-	local p = Platform
-	local d_before = p.developer
-	p.developer = true
-
-	-- fixes UpdateInterface nil value in editor mode
-	editor.LoadPlaceObjConfig()
-	-- needed for HashLogToTable(), SM was planning to have multiple cities (or from a past game from this engine)?
-	if not rawget(_G,"g_Cities") then
-		GlobalVar("g_Cities",{})
-	end
-	-- editor wants a table
-	GlobalVar("g_revision_map",{})
-	-- stops some log spam in editor (function doesn't exist in SM)
-	UpdateMapRevision = rawget(_G,"UpdateMapRevision") or empty_func
-	AsyncGetSourceInfo = rawget(_G,"AsyncGetSourceInfo") or empty_func
-
-	p.developer = d_before
-end -- do
-
 do -- non-class obj funcs
 	-- stops the help webpage from showing up every single time
 	SaveOrigFunc("GedOpHelpMod")